Password change page: 'Previous password' -> 'Current Password'
This is in-line with what I see the 'give your password to change your
password' field being on most websites. I also changed the POST
parameter name and ran the `black` formatter over the file.

I had to change one line to move the ascii decoding into another
statement, because (due to line length), black was clobbering the file
in a really ugly way :(